feat: add guest access functionality for quizzes and shared links
- Created migration to add guest columns to quiz_attempts table.
- Updated migration to allow null user_id in quiz_attempts table.
- Implemented a new Table component for better UI consistency.
- Developed Shared Links management page with CRUD operations and statistics.
- Added statistics page for shared links, displaying usage data and session details.
- Implemented GuestQuiz component for guest users to take quizzes without authentication.
- Defined new routes for guest access to quizzes and shared links management.

+++ config/wecan.php
@@ -1,9 +1,9 @@
 <?php
 
 return [
-    'access_price'     => env('WECAN_ACCESS_PRICE', 5000),
+    'access_price' => env('WECAN_ACCESS_PRICE', 5000),
     'access_price_usd' => env('WECAN_ACCESS_PRICE_USD', 5),
-    'quiz_questions'   => env('QUIZ_QUESTION_COUNT', 20),
-    'quiz_duration'    => env('QUIZ_DURATION_MINUTES', 20),
-    'pass_score'       => env('QUIZ_PASS_SCORE', 70),
+    'quiz_questions' => (int) env('QUIZ_QUESTION_COUNT', 20),
+    'quiz_duration' => (int) env('QUIZ_DURATION_MINUTES', 20),
+    'pass_score' => env('QUIZ_PASS_SCORE', 70),
 ];
